server/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
flatbuffers==1.12
fonttools==4.33.3
gast==0.3.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.11.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.5.2
mediapipe==0.8.7.3
numpy==1.19.5
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.3
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.15.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.9
websockets==10.3
Werkzeug==2.1.2
wrapt==1.12.1
zipp==3.8.0
